22/03/2022 Tiến độ hoàn thành dự án 80%

+++ WebShopData/Models/ViewCtdonHang.cs
@@ -17,6 +17,6 @@ namespace WebShopData.Models
         public string TenMauSac { get; set; }
         public string ImageUrl { get; set; }
         public long Gia { get; set; }
-        public long SoLuong { get; set; }
+        public int SoLuong { get; set; }
     }
 }
